refactor pages and routing
