+++ public/index.html
@@ -1,30 +1,30 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <title>Seth Morenos</title>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.png" />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/apple-touch-icon.png" />
-    <meta name="author" content="Seth Morenos" />
-    <meta
-        name="description" 
-        content="Hello! I'm Seth, a designer and Engineering student at the University of Waterloo."
-    />
-    <meta property="og:type" content="website" />
-    <meta property="og:url" content="https://sethvm.com/" />
-    <meta property="og:title" content="Seth Morenos - Portfolio" />
-    <meta property="og:image" content="%PUBLIC_URL%/preview.png" />
-    <meta 
-        property="og:description" 
-        content="Designer, Engineering Student" 
-    />
-    <meta property="og:site_name" content="Seth Morenos" />
-    <meta name="theme-color" content="#000000" />
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-  </body>
+    <head>
+        <title>Seth Morenos</title>
+        <!-- meta tags -->
+        <meta charset="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <meta name="author" content="Seth Morenos">
+        <meta name="theme-color" content="#000000" />
+        <!-- open graph metas-->
+        <meta property="og:title" content="Seth Morenos - Portfolio" />
+        <meta property="og:type" content="website" />
+        <meta property="og:site_name" content="sethvm.com" />
+        <meta property="og:description" content="Designer || Engineering Student">
+        <meta property="og:image" content="%PUBLIC_URL%/preview.png" />
+        <meta property="og:url" content="https://sethvm.com/" />
+        <!-- twitter graph metas-->
+        <meta name="twitter:title" content="European Travel Destinations ">
+        <meta name="twitter:description" content="Designer || Engineering Student">
+        <meta name="twitter:image" content="%PUBLIC_URL%/preview.png">
+        <!-- icon and preview -->
+        <link rel="icon" href="%PUBLIC_URL%/favicon.png" />
+        <link rel="apple-touch-icon" href="%PUBLIC_URL%/apple-touch-icon.png" />
+        <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    </head>
+    <body>
+        <noscript>You need to enable JavaScript to run this app.</noscript>
+        <div id="root"></div>
+    </body>
 </html>
