--- src/App.css
@@ -1,4 +1,0 @@
-.page_content {
-    max-width: 1920px;
-    margin: auto;
-}
