+++ src/pages/Home.js
@@ -1,15 +1,23 @@
 import React from 'react';
 
-import './Home.css';
+import '../components/Home.css';
+
+import HomeHeader from '../components/Header/HomeHeader';
+import Carousel from '../components/Carousel/Carousel';
+import SideBar from '../components/SideBar/SideBar';
+import BlueBlock from '../components/BlueBlock/BlueBlock';
+import HomeFooter from '../components/Footer/HomeFooter';
 
 export default function Home() {
+    
     return (
         <>
+        <HomeHeader />
         <div className='intro_sm animated fadeIn'>
             <h1><strong>DESIGNER</strong></h1>
             <h6><strong>ENGINEERING STUDENT</strong></h6>
             <h6><strong>UNIVERSITY OF WATERLOO</strong></h6>
-        </div>
+        </div> {/* /intro-sm */}
         <div className='intro_md animated fadeIn'>
             <h1><strong>Hello! I'm Seth.</strong></h1>
             <p>
@@ -20,7 +28,7 @@ export default function Home() {
                 <br />with a passion for creating products that leave a positive impact
                 <br />in the world around me.
             </p>
-        </div>
+        </div> {/* /intro-md */}
         <div className='intro_lg animated fadeIn'>
             <h1>
                 <strong>Hello! I'm Seth.</strong>
@@ -40,7 +48,11 @@ export default function Home() {
                 <br />Previously a <strong>UX and Web Designer</strong>
                 <br />at the <strong>Ontario Public Service</strong>.
             </p>
-        </div>
+        </div> {/* /intro-lg */}
+        <Carousel />
+        <SideBar />
+        <BlueBlock />
+        <HomeFooter />
         </>
     );
 }
