+++ src/pages/PageContainer.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function PageContainer(props) {
+    return (
+        <div className='project_content'>
+            {props.children}
+        </div>
+    );
+}
